wait for result to be printed before exiting
Measurements are read from a channel and, in silent mode, the
only output is printed after all measurements have been read.
That is when the channel provided to fast.Measure() has been
closed at the end of fast.Measure() and drained by the
goroutine reading from the measurements.

main() must wait for that goroutine that eventually prints the
test result before exiting.

Signed-off-by: David Verbeiren <david.verbeiren@tessares.net>